% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{is.forecast}
\alias{is.forecast}
\title{Is an object a particular forecast type?}
\usage{
is.forecast(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
Returns true if the forecast object is of a particular type
}
